backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
python-dotenv
python-multipart
pydantic
tqdm
gunicorn
